Removed donation drive.
Not even $1 was donated and no one was true to their word.
I'm glad my motivation to code doesn't rely on this community.
If the most trusted developer in the community can't raise money, then
the only ones receiving money are those who wish to exploit users and
the game.